Fix HealthSystem
onCollisionEnter buscando en el padre

+++ Assets/Test/TestTarget.cs
@@ -5,14 +5,14 @@ public class TestTarget : MonoBehaviour, IDamageable
     [SerializeField] private int maxHealth = 30;
     [SerializeField] private int scoreValue = 10;
 
-    private int _currentHealth;
+    private float _currentHealth;
 
     private void Awake()
     {
         _currentHealth = maxHealth;
     }
 
-    public void TakeDamage(int amount)
+    public void TakeDamage(float amount)
     {
         _currentHealth -= amount;
         Debug.Log($"{name} recibio {amount} de damage. Vida restante: {_currentHealth}");
